requirements.txt
Flask
Flask-SQLAlchemy
python-dotenv
Flask-WTF
werkzeug